Create Quoters and encapsulate business logic
