+++ src/routers/alpha-router/quoters/model/results/get-quotes-result.ts
@@ -1,0 +1,7 @@
+import { RouteWithValidQuote } from '../../../entities';
+import { CandidatePoolsBySelectionCriteria } from '../../../functions/get-candidate-pools';
+
+export interface GetQuotesResult {
+  routesWithValidQuotes: RouteWithValidQuote[];
+  candidatePools?: CandidatePoolsBySelectionCriteria;
+}
