+++ src/routers/alpha-router/quoters/model/results/get-routes-result.ts
@@ -1,0 +1,7 @@
+import { MixedRoute, V2Route, V3Route } from '../../../../router';
+import { CandidatePoolsBySelectionCriteria } from '../../../functions/get-candidate-pools';
+
+export interface GetRoutesResult<Route extends V2Route | V3Route | MixedRoute> {
+  routes: Route[];
+  candidatePools: CandidatePoolsBySelectionCriteria;
+}
